Remove UserJobListView
Reason: the home page will show a login option and a welcome graphic and
nothing more. the logic already built in means it will show that unless
user is logged in where it will show the users jobs. This might not be a
good idea to have two big pages in one page so this might change later.