Move to PostgreSQL, update settings.py accordingly
Moved to PostgreSQL and experienced a lot of pain in the setup process.
I am now familiar with `psql`, `pg_ctl`, and `postgres` as command line options. Testing on a different system will require significant user intervention so we'll have to write some kind of setup script to make that smoother. I have written down some rough ideas in the README just in case.