Add new job model and form
Create a new model based on the backend metagenome config file. Also
remove or comment out the django-tables2 references in the code. DT2 is
a nice interface to make tables but we only need one table as far as I
can tell so perhaps it's best to not overcomplicate things right now.