Add both mouse wheel and brush zoom to same script
Add the ability to use mouse wheel to fine tune the zoom in
functionality. If the user wants to zoom out after zooming in too far,
there is no alternative but to reset the zoom and try again. This may be
frustrating for interacting multiple times with the same plot. The
pan-zoom functionality offered by the mouse wheel is and good amendment
to this problem. The biggest problem is that the zoom state is changed
by the two states independently. The brush and mouse wheel need to share
the zoom states in order to function smoothly. This is still a problem
and will likely require a lot more understanding of how `d3.zoom` works.