Use WebGL to visualize 700k points
- Use [`d3fc`](https://d3fc.io/) to visualize 700k points using their
[diamond dataset visualization](https://bl.ocks.org/OliverForeman/f8bc8d69b7df041ada60cc86952c6fee). This was successful however, it requires some understanding of WebGL just to visualize. Interactivity likely requires further study of the WebGL language, something I am unwilling to gain at this point. Abandonded.

- Use [x3dom](https://www.x3dom.org/). Supposedly gives you greater
flexibility and control over WebGL but requires a heavy tutorial to get
started with basics. Abandoned (perhaps hastily).

- Use [plotly.js](https://plotly.com/javascript/webgl-vs-svg/) with
WebGL. Based on my limited understanding of `plotly`'s methods, it
requires pushing all points to be plotted into an array and then
plotting a new graph with an argument containing that array. This is not
useful for 1M points.

- Additional consideration: trying to visualize 1M points as a scatter
plot may not be the best idea. Instead, there might be a better
visualization technique such as subsampling or binning. Consider that in a
1000x1000 px plot area, there are only a million pixels to visualize
these points. Due to the nature of the data, there will be some
[overplotting](https://www.data-to-viz.com/caveat/overplotting.html). In
order to reduce this, I have decided to use
[hexbin](https://github.com/d3/d3-hexbin) to bin the data such that when
the user clicks on one bin, it zooms in and displays either more bins or
the points within that bin. The number of points in a bin is dependent
on the total number of points I can visualize using `svg` which is
around 1-2k based on [the OpenVis conference talk by Dominikus
Baur](https://youtu.be/dlZvL7Ei0C0). Given this limit, I can partition
my points into `total points / limit of visualization` bins. Within a
given bin, this must be repeated unless the `total points in bin <
visualization limit`.